change destroy

+++ app/controllers/items_controller.rb
@@ -44,8 +44,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    if user_signed_in? && item.user_id == current_user.id
-      item.destroy
+    if user_signed_in? && item.user_id == current_user.id && item.destroy
       redirect_to root_path
     else
       render "show"
